requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.6
Django==4.0.2
django-active-link==0.1.8
django-appconf==1.0.5
django-classy-tags==3.0.1
django-crispy-forms==1.14.0
django-filter==21.1
lxml==4.7.1
mysqlclient==2.1.0
Pillow==9.0.1
sqlparse==0.4.2
